requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
Cerberus==1.3.7
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
decorator==5.2.1
dj-database-url==3.0.1
Django==5.2.4
django-allauth==65.10.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-health-check==3.20.0
django-prometheus==2.4.1
django-redis==6.0.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
executing==2.2.0
flower==2.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
idna==3.10
inflection==0.5.1
ipdb==0.13.13
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kafka-python==2.2.15
kombu==5.5.4
marshmallow==4.0.0
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
structlog==25.4.0
text-unidecode==1.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
